+++ Source/Public/G1M/G1MM.h
@@ -8,9 +8,9 @@ struct G1MM
 {
 	uint32_t matrixCount;
 	std::vector<RichMat44> matrices;
-	G1MM(BYTE* buffer, size_t startOffset)
+	G1MM(BYTE* buffer, uint32_t startOffset)
 	{
-		size_t offset = startOffset;
+		uint32_t offset = startOffset;
 		GResourceHeader sectionHeader = reinterpret_cast<GResourceHeader<bBigEndian>*>(buffer + offset);
 		offset = startOffset + 12;
 		matrixCount = *reinterpret_cast<uint32_t*>(buffer+offset);
